requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
asyncio==3.4.3
aiohttp==3.9.1
beautifulsoup4==4.12.2
playwright==1.40.0
Pillow==10.1.0
numpy==1.24.3
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.2.2
chromadb==0.4.22
langchain==0.1.0
accelerate==0.25.0
bitsandbytes==0.41.3
einops==0.7.0
rank-bm25==0.2.2
scikit-learn==1.3.2
psutil==5.9.6
GPUtil==1.4.0
pytest==7.4.3
pytest-asyncio==0.21.1
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9